pyproject.toml
[project]
name = "ai-agents"
version = "0.1.0"
description = ""
authors = [
    {name = "František Šitner",email = "frantisek.sitner@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai (>=1.100.2,<2.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "litellm (>=1.77.5,<2.0.0)",
    "devtools (>=0.12.2,<0.13.0)",
    "langgraph (>=0.6.7,<0.7.0)",
    "langchain-mcp-adapters (>=0.1.10,<0.2.0)",
    "langchain-ollama (>=0.3.8,<0.4.0)"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
